Window-based Randomized SVD on data streamed from GenomicDataStream

Window-based Randomized SVD on data streamed from GenomicDataStream

Usage

PCAstream(gds, k, p = 7, s = 20, B = 64, threads = 4, verbose = TRUE)

Arguments

gds

GenomicDataStream

k

integer;
specifies the target rank of the low-rank decomposition. \(k\) should satisfy \(k << min(m,n)\).

p

integer, optional;
number of additional power iterations (by default \(p=7\)).

s

integer, optional;
oversampling parameter (by default \(s=20\)).

B

integer, optional;
number of windows (by default \(B=64\)).

threads

integer, optional;
number of threads (by default \(threads=4\)). Set to min(threads, floor(nrow(chunks) / B))

verbose

string, optional;
if TRUE (default) print details

Value

PCAstream returns a list containing the following three components:

d

array_like;
singular values; vector of length \((k)\).

u

array_like;
left singular vectors; \((m, k)\) or \((m, nu)\) dimensional array.

v

array_like;
right singular vectors; \((n, k)\) or \((n, nv)\) dimensional array.

Details

PCAstream implements the window-based Randomized SVD proposed by Li et al. (2024)

Note

The singular vectors are not unique and only defined up to sign. If a left singular vector has its sign changed, changing the sign of the corresponding right vector gives an equivalent decomposition.

References

Li, Z., Meisner, J., & Albrechtsen, A. (2023). Fast and accurate out-of-core PCA framework for large scale biobank data. Genome Research, 33(9), 1599-1608. doi:10.1101/gr.277525.122 .

Author

Zilong Li zilong.dk@gmail.com

Examples

file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")

obj <- GenomicDataStream(file, "DS", chunkSize = 3)

res <- PCAstream(obj, k=2, B=2)
#>  # features: 10 
#>  # chunks: 4 
#>  # threads: 2 
#>  B: 2 
#>  k: 2 

str(res)
#> List of 4
#>  $ d     : num [1:2] 9.81 9.14
#>  $ u     : num [1:10, 1:2] -0.5392 -0.0777 -0.036 -0.0579 -0.0378 ...
#>  $ v     : num [1:60, 1:2] -0.2195 -0.1519 0.0101 -0.0741 -0.0271 ...
#>  $ timing: Named num [1:13] 0 145714 1589037 1698999 3248102 ...
#>   ..- attr(*, "names")= chr [1:13] "init params" "init params" "init MultiGenomicStreamProcessor" "init Eigen::MatrixXd for pcaone" ...