Fit series of linear regression models

Fit series of linear regression models

Usage

# S4 method for class 'ANY,ANY,GenomicDataStream'
lmFitFeatures(
  y,
  design,
  data,
  weights,
  detail = 1,
  preprojection = TRUE,
  nthreads = 1,
  ...
)

Arguments

y

response vector

design

design matrix, mat or sp_mat

data

matrix or GenomicDataStream with additional features to be fit one at a time

weights

sample-level weights

detail

return model with specified level of detail. LOW (beta, se, sigSq, rdf), MEDIUM (vcov), HIGH (residuals), MOST (hatvalues)

preprojection

default true. Use preproject of design matrix to accelerate calculations

nthreads

number of threads. Each model is fit in serial, analysis is parallelized across features

...

other args

Examples

# create response, design and weights
y <- seq(60)
X_design <- matrix(1, 60, 1)
w <- rep(1, 60)

# VCF file
file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")

# Read data into R
# then run lmFitFeatures()
gds <- GenomicDataStream(file, "DS", initialize = TRUE)
dat <- getNextChunk(gds)
X_features <- dat$X

res1 <- lmFitFeatures(y, X_design, X_features, w)

# Data stays at C++ level
# then run lmFitFeatures()
gds <- GenomicDataStream(file, "DS")

res2 <- lmFitFeatures(y, X_design, gds, w)